refactor: LightningTask (seperate LightningModule from task functionality)
