+++ embeddings/pipeline/lightning_sequence_labeling.py
@@ -15,8 +15,8 @@ from embeddings.evaluator.sequence_labeling_evaluator import (
 )
 from embeddings.model.lightning_model import LightningModel
 from embeddings.pipeline.lightning_pipeline import LightningPipeline
-from embeddings.task.lightning_task.sequence_labeling import SequenceLabeling
 from embeddings.utils.json_dict_persister import JsonPersister
+from embeddings.task.lightning_task.sequence_labeling import SequenceLabelingTask
 from embeddings.utils.utils import initialize_kwargs
 
 
@@ -78,17 +78,17 @@ class LightningSequenceLabelingPipeline(
             load_dataset_kwargs=load_dataset_kwargs,
             **self.datamodule_kwargs
         )
-        trainer = pl.Trainer(default_root_dir=output_path, **self.task_train_kwargs)
-
-        task = SequenceLabeling(
+        task = SequenceLabelingTask(
             model_name_or_path=embedding_name,
+            output_path=output_path,
             train_batch_size=train_batch_size,
             eval_batch_size=eval_batch_size,
             finetune_last_n_layers=finetune_last_n_layers,
-            config_kwargs=self.model_config_kwargs,
+            model_config_kwargs=self.model_config_kwargs,
             task_model_kwargs=self.task_model_kwargs,
+            task_train_kwargs=self.task_train_kwargs,
         )
-        model = LightningModel(trainer=trainer, task=task, predict_subset=predict_subset)
+        model = LightningModel(task=task, predict_subset=predict_subset)
         evaluator = SequenceLabelingEvaluator(
             evaluation_mode=evaluation_mode, tagging_scheme=tagging_scheme
         ).persisting(JsonPersister(path=output_path.joinpath(evaluation_filename)))
